+++ PhojTree_Check/DelPanj/TreeMaker/test/runMC/treemaker425MC_3_1_cfg.py
@@ -128,7 +128,7 @@ process.zeefil.subLeadElecPset_.ptx= cms.double(20)
 
 from DelPanj.TreeMaker.photonFilter_cff import *
 process.photonfil = photonFilter
-process.photonfil.GammaPtMinPY=cms.double(85)
+process.photonfil.GammaPtMinPY=cms.double(40)
 
 #process.p = cms.Path(process.zeefil*process.demo)
 process.p = cms.Path(
